+++ test/index.html
@@ -7,6 +7,7 @@
 	<script src="../dist/qunit.js"></script>
 	<script src="test.js"></script>
 	<script src="dump.js"></script>
+	<script src="modules.js"></script>
 	<script src="deepEqual.js"></script>
 	<script src="globals.js"></script>
 	<script src="swarminject.js"></script>
